Serializers devided in different files, LOCAL_HOST added to .env file
