+++ django/entrypoint.sh
@@ -11,6 +11,8 @@ echo "PostgreSQL started"
 #  TODO:Write better code cuz if theres no migrations folder this wont work
 # Check if migrations have been applied for the dsp app
 #if python src/manage.py showmigrations | grep -q "(no migrations)"; then
+
+# Check if there is not migrations  folder apply migrations and fill the categories and subcategories
 if [ ! -d "src/tdsp/dsp/migrations" ]; then
     echo "Migrations for dsp app not applied. Running makemigrations and migrate..."
     python src/manage.py makemigrations dsp
@@ -20,7 +22,6 @@ if [ ! -d "src/tdsp/dsp/migrations" ]; then
     python src/manage.py createsuperuser --email=admin@admin.com --noinput
     echo "Superuser created!"
 else
-    ls
     echo "Migrations for dsp app already applied. Skipping makemigrations and migrate."
 fi
