+++ django/src/tdsp/api/campaign_api.py
@@ -1,9 +1,8 @@
-from django.shortcuts import get_object_or_404
-from rest_framework import viewsets, status
-from rest_framework.response import Response
-from ..serializers.serializers import CampaignSerializer
+from rest_framework import viewsets
+
 from ..dsp.models.campaign_model import CampaignModel
-from ..dsp.models.game_config_model import ConfigModel
+
+from ..serializers.campaign_serializer import CampaignSerializer
 
 
 class CampaignViewSet(viewsets.ModelViewSet):
