+++ django/src/tdsp/api/user_group_api.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import User
 
 from rest_framework import viewsets
 
-from ..serializers.serializers import UserSerializer, UserCreateSerializer
+from ..serializers.user_serializers import UserSerializer, UserCreateSerializer
 from ..permissions.permissions import IsAdminOrReadOnly
