+++ django/src/tdsp/serializers/user_serializers.py
@@ -1,0 +1,31 @@
+from django.contrib.auth.models import User
+
+from rest_framework import serializers
+
+
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = User
+        fields = ['id', 'url', 'username', 'email', 'is_staff']
+
+    def create(self, validated_data):
+        user = User.objects.create_user(**validated_data)
+        return user
+
+
+class UserCreateSerializer(UserSerializer):
+    password = serializers.CharField(write_only=True)
+    is_staff = serializers.BooleanField(default=False, write_only=True)
+
+    def create(self, validated_data):
+        user = User.objects.create_user(
+            username=validated_data['username'],
+            email=validated_data['email'],
+            password=validated_data['password'],
+            is_staff=validated_data['is_staff']
+        )
+        return user
+
+    class Meta:
+        model = User
+        fields = ['id', 'username', 'email', 'password', 'is_staff']
