+++ django/src/tdsp/tools/image_server_tools.py
@@ -97,11 +97,8 @@ def save_image_to_minio(base64_image):
                           'ContentType': content_type,
                           'ContentDisposition': 'inline'
                       })
-    if env.bool('DEBUG', default=False):
-        # Generate a public URL for the image
-        url = f'http://localhost:9000/{bucket_name}/{name}'
-    else:
-        url = f'http://localhost/{bucket_name}/{name}'
+
+    url = f"http://{env.str('LOCAL_HOST')}/{bucket_name}/{name}"
 
     # Return the URL to the Django application
     return url
