+++ env.example
@@ -2,6 +2,7 @@ SECRET_KEY=some_secret_key
 DEBUG=True/False
 DJANGO_SUPERUSER_USERNAME=admin
 DJANGO_SUPERUSER_PASSWORD=admin
+LOCAL_HOST = machine_local_ip/localhost
 
 POSTGRES_NAME=pg_name
 POSTGRES_DB=pg_db_name
